src/auth_server/pyproject.toml
[tool.poetry]
name = "auth-server"
version = "0.1.0"
description = ""
authors = ["Hess <hess90wang@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.4"
django = "^5.1.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
grpcio = "^1.66.1"
django-environ = "^0.11.2"
drf-yasg = "^1.21.7"
mysqlclient = "^2.2.4"
protobuf = "^5.28.0"
grpcio-tools = "^1.66.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
